Add csv_content parameter to Sheet class
The constructor of Sheet class has been updated to accept a new parameter csv_content. This parameter specifies the type of content that the CSV file holds. Also, the methods in the class have been updated to handle data accordingly, based on the csv_content. This allows more flexibility and error handling when dealing with different types of CSV data.